backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.10
alembic==1.16.5
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt<4
python-multipart==0.0.20
pydantic-settings==2.10.1
email-validator==2.2.0
httpx==0.28.1
beautifulsoup4==4.13.4
redis==6.4.0
apscheduler==3.11.0
openpyxl==3.1.5
google-api-python-client==2.181.0
google-auth==2.40.3
